[frontend] Centre viewport at medoid of player's cells
