+++ salz-web/lib/game/jwt.js
@@ -1,0 +1,14 @@
+export function parseJWT(token) {
+  const base64Payload = token.split('.')[1];
+  const base64 = base64Payload.replace(/-/g, '+').replace(/_/g, '-');
+  const jsonPayload = decodeURIComponent(
+    atob(base64)
+      .split('')
+      .map((c) => {
+        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
+      })
+      .join('')
+  );
+
+  return JSON.parse(jsonPayload);
+}
